feat: add concurrent LRU cache for DID documents
